iAssembler: Replace parameter for assembler function and add format string
The properties for the assembler programm are stored within an configuration
file. This file is in Windows-Ini format and can be parsed to a property
tree via boost library functions. This property map is used within the
assembler function of a parsed object to read out necessary configuration
data like VCGRA features or assembler environment variables.

The stored format string is used to create machine code readable by the
VCGRA management unit.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>